Rename model_metadata module to modelmetadata.
